コンポーネント化しよう
