+++ src/components/Footer.tsx
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <footer className="bg-blue-500 text-white text-sm py-4 text-center mt-auto">
+      &copy; 2025 full stack web development
+    </footer>
+  );
+}
