+++ src/components/Header.tsx
@@ -1,0 +1,7 @@
+export default function Header() {
+  return (
+    <header className="bg-blue-500 text-white text-lg font-bold py-4 px-6 flex justify-between items-center">
+      在庫管理システム
+    </header>
+  );
+}
